Add `StatefulModel` (#413)
